requirements.txt
streamlit
PyMuPDF
sentence-transformers
faiss-cpu
tiktoken
requests
deep-translator
gTTS
python-dotenv